feature: got tests to run without erorrs, fixed clippy
