+++ tests/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod dkg;
 pub mod mocks;
 pub mod protocol;
